test_running/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pyyaml==6.0.1
pydantic==2.5.0
jinja2==3.1.2
asyncio
prometheus_client==0.19.0
requests==2.31.0 